+++ NEWS.md
@@ -11,6 +11,7 @@
 - `backendInitialize()` with defined parameter `data` drops core spectra
   variables that contain only missing values.
 - Add `$<-` method for `MsBackendPy`.
+- Add `peaksData()<-` for `MsBackendPy`.
 
 ## Changes in 0.99.11